Prepare #935: extract code from existing o:socket impl into code reusable for new o:sse
